Split up the GUI file into a Controller and a View file.
